[FIX] ws doc: fix DOM order for copy/paste usability
The toggle button for the setup code in code examples of the webservice
page was mistakenly inserted between the setup code and the "actual"
code in the DOM. Trying to select the whole snippet (including setup)
would thus also select the text of the button and copy it to the
clipboard, breaking the copied code and confusing users.

Fix DOM order so selecting setup and actual code does not include the
text of the toggle button.