Declare PyChart package name in requirements.txt
While calling `pip install -r requirements.txt`, the PyChart tarball is
always downloaded even if `pip` has a download cache configured (which
it does by default in later versions).

This is because `pip` has no way of knowing which package the PyChart
tarball contains before downloading and running its `setup.py`.

By declaring the PyChart "egg" in `requirements.txt` the tarball is no
longer downloaded if it's already in the download cache.