[FIX] search view: restore the `disable_custom_filters` flag effect
This search view option was properly set but not
observed by the search view itself.

This caused all related links and smart buttons
to open a list of records that was wrongly filtered,
as soon as a default filter was applicable to the
target action (e.g. with default filters gloablly
set per model, not per action).

This regression was introduced at refactoring
commit d8d6311b50d0d3446cab3f3386ad672ac486bab2