[FIX] google_spreadsheet: 'add to spreadsheet' in Favorites
The option 'add to spreadsheet' works under the hypothesis that the
current action is a ViewManager. It crashes in other contexts (e.g. in
Messaging, mail box), as it tries to retrieve the list_view id of the
ViewManager.

We now only add the option in the Favorite menu if the current action is
a ViewManager.