[IMP] Web, Web_diagram: better integration of the Diagram view
The 'New node' button is inserted in the ControlPanel. Breadcrumbs
are now handled as well.

The 'multi_record' flag has been added to the views to distinguish views
displaying several records, like the List view, from views displaying
a single record like the Form view and the Diagram view. This flag is
used by the ViewManager to manage its view stack. Fixes the breadcrumb
bug when switching between Form view and Diagram view (the last element
of the view stack is now replaced by the new one in that case).